% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptivesplines.R
\name{.periodic_adaptive_spline}
\alias{.periodic_adaptive_spline}
\title{Title}
\usage{
.periodic_adaptive_spline(
  x,
  y,
  period,
  nknots,
  knots = NULL,
  order = 4,
  lambda = 0.1,
  maxiter = 20,
  precision = 0.001,
  threshold = 0.98
)
}
\arguments{
\item{threshold}{}
}
\description{
Title
}
