% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptivesplines.R
\name{.periodic_adaptive_splines}
\alias{.periodic_adaptive_splines}
\title{Title}
\usage{
.periodic_adaptive_splines(
  x,
  y,
  period,
  nknots,
  knots = NULL,
  order = 4,
  maxiter = 20,
  precision = 0.001,
  threshold = 0.98,
  lambda0 = 0,
  lambda1 = 10,
  dlambda = 0.025,
  minKnots = 0,
  criterion = "AIC"
)
}
\arguments{
\item{criterion}{}
}
\description{
Title
}
