% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjd3xjars-package.R
\docType{package}
\name{rjd3xjars-package}
\alias{rjd3xjars}
\alias{rjd3xjars-package}
\title{rjd3xjars: Interface to 'JDemetra+ 3.x' Time Series Analysis Software}
\description{
R Interface to 'JDemetra+ 3.x' (\url{https://github.com/jdemetra}) time series analysis software. It provides functions allowing to decompose a time series, including high-frequency data with multiple periodicities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rjdverse/rjd3xjars}
  \item Report bugs at \url{https://github.com/rjdverse/rjd3xjars/issues}
}

}
\author{
\strong{Maintainer}: Tanguy Barthelemy \email{timeserieswithjdemetraandr@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Tanguy Barthelemy \email{timeserieswithjdemetraandr@gmail.com} [copyright holder]
  \item Jean Palate \email{palatejean@gmail.com}
}

}
\keyword{internal}
