% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdarima.R
\name{ltdarima_decomposition}
\alias{ltdarima_decomposition}
\title{Estimation by means of the Kalman smoother of the canonical decomposition
of linear time-dependent SARIMA models}
\usage{
ltdarima_decomposition(data, regular, seasonal, p0, p1, var1 = 1, se = FALSE)
}
\arguments{
\item{data}{The series (a ts).}

\item{regular}{The orders of the regular part of the SARIMA models (p,d,q)}

\item{seasonal}{The orders of the seasonal part of the SARIMA models (bp,bd,bq). NULL is (0,0,0)}

\item{p0}{The arima parameters at the beginning of the series}

\item{p1}{The arima parameters at the end of the series}

\item{var1}{The (unscaled) variance of the innovations at the end of the series (1 at the beginning of the series)}

\item{se}{Specifies if the standard deviations of the components are computed}
}
\value{
A mts object is returned, with the trend, the seasonal, the irregular components and - if requested - their standard deviations
}
\description{
Estimation by means of the Kalman smoother of the canonical decomposition
of linear time-dependent SARIMA models
}
\details{
The arima parameters are organized in the following order: phi, bphi, theta, btheta
}
\examples{
\dontshow{if (rjd3toolkit::get_java_version() >= rjd3toolkit::minimal_java_version) withAutoprint(\{ # examplesIf}
s <- rjd3toolkit::Retail$RetailSalesTotal

q <- rjd3advsajars::ltdarima_estimation(
    data = s,
    regular = c(0, 1, 1),
    seasonal = c(0, 1, 1),
    fixed_var = FALSE,
    eps = 1e-15,
    parametrization = "mean_delta"
)
sa <- rjd3advsajars::ltdarima_decomposition(
    s,
    c(0, 1, 1),
    c(0, 1, 1),
    q$final$model$parima_0,
    q$final$model$parima_1,
    var1 = q$final$model$parameters[5],
    se = TRUE
)
ts.plot(
    ts.union(
        s,
        sa$components[, 1],
        sa$components[, 1] + sa$components_stdev[, 1]
    ),
    col = c('gray', 'blue', 'magenta')
)
\dontshow{\}) # examplesIf}
}
