% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdarima.R
\name{tdarima_decomposition}
\alias{tdarima_decomposition}
\title{Estimation by means of the Kalman smoother of a time-dependent canonical decomposition (generic solution)}
\usage{
tdarima_decomposition(data, regular, seasonal, parameters, se = FALSE)
}
\arguments{
\item{data}{The series (a ts).}

\item{regular}{The orders of the regular part of the SARIMA models (p,d,q)}

\item{seasonal}{The orders of the seasonal part of the SARIMA models (bp,bd,bq). NULL is (0,0,0)}

\item{parameters}{The time dependent parameters. The matrix contains for each column
all the parameters of the model, with the columns corresponding to an observation. So,
the number of columns is identical to the length of the series and the number of rows corresponds
to the arima parameters (regular + seasonal), with eventually the innovation variances}

\item{se}{Specifies if the standard deviations of the components are computed}
}
\value{
A mts object is returned, with the trend, the seasonal, the irregular components and - if requested - their standard deviations
}
\description{
Estimation by means of the Kalman smoother of a time-dependent canonical decomposition (generic solution)
}
\details{
The arima parameters are organized in the following order: phi, bphi, theta, btheta
}
\examples{
\dontshow{if (rjd3toolkit::get_java_version() >= rjd3toolkit::minimal_java_version) withAutoprint(\{ # examplesIf}
s <- rjd3toolkit::ABS$X0.2.09.10.M
n <- length(s)
r <- runif(n * 2, -1, -.6)
p <- matrix(r, nrow = 2, ncol = n)
p <- rbind(p, runif(n))
sa <- rjd3advsajars::tdarima_decomposition(
    s,
    c(0, 1, 1),
    c(0, 1, 1),
    parameter = p,
    se = TRUE
)
ts.plot(
    ts.union(s, sa$components[, 1], sa$components[, 1] + sa$components[, 3]),
    col = c('gray', 'blue', 'magenta')
)
ts.plot(sa$components_stdev, col = c('red', 'blue', 'magenta'))
\dontshow{\}) # examplesIf}
}
